requirements.txt
fastapi
uvicorn[standard]
torch
albumentations
numpy
Pillow
rasterio
jinja2
tqdm